# [dbnode] Improve m3tsz decoding performance