--- a/src/dbnode/encoding/m3tsz/timestamp_iterator.go
+++ b/src/dbnode/encoding/m3tsz/timestamp_iterator.go
@@ -22,12 +22,18 @@ package m3tsz
 
 import (
 	"encoding/binary"
-	"fmt"
+	"errors"
 	"time"
 
 	"github.com/m3db/m3/src/dbnode/encoding"
 	"github.com/m3db/m3/src/dbnode/ts"
 	xtime "github.com/m3db/m3/src/x/time"
+)
+
+var (
+	errNoTimeSchemaForUnit        = errors.New("time encoding scheme doesn't exist for unit")
+	errUnexpectedAnnotationLength = errors.New("expected annotation length to be >= 0")
+	errAnnotationTooFewBytes      = errors.New("expected to read annotation bytes, but got end of stream")
 )
 
 // TimestampIterator encapsulates all the state required for iterating over
@@ -37,11 +43,12 @@ type TimestampIterator struct {
 	PrevTimeDelta time.Duration
 	PrevAnt       ts.Annotation
 
-	TimeUnit xtime.Unit
+	TimeUnit        xtime.Unit
+	defaultTimeUnit xtime.Unit
 
-	Opts encoding.Options
-
-	markerEncodingScheme encoding.MarkerEncodingScheme
+	markerEncodingScheme *encoding.MarkerEncodingScheme
+	timeEncodingSchemes  encoding.TimeEncodingSchemes
+	timeEncodingScheme   *encoding.TimeEncodingScheme
 
 	TimeUnitChanged bool
 	Done            bool
@@ -59,11 +66,12 @@ type TimestampIterator struct {
 func NewTimestampIterator(opts encoding.Options, skipMarkers bool) TimestampIterator {
 	mes := opts.MarkerEncodingScheme()
 	return TimestampIterator{
-		Opts:                 opts,
+		defaultTimeUnit:      opts.DefaultTimeUnit(),
 		SkipMarkers:          skipMarkers,
 		numValueBits:         uint8(mes.NumValueBits()),
 		numBits:              uint8(mes.NumOpcodeBits() + mes.NumValueBits()),
 		markerEncodingScheme: mes,
+		timeEncodingSchemes:  opts.TimeEncodingSchemes(),
 	}
 }
 
@@ -73,14 +81,22 @@ func (it *TimestampIterator) ReadTimestamp(stream *encoding.IStream) (bool, bool
 
 	var (
 		first = false
+		dod   time.Duration
 		err   error
 	)
-	if it.PrevTime == 0 {
+
+	if it.PrevTime != 0 {
+		// inlined readNextTimestamp
+		dod, err = it.readMarkerOrDeltaOfDelta(stream)
+		if err == nil {
+			it.PrevTimeDelta += dod
+			it.PrevTime += xtime.UnixNano(it.PrevTimeDelta)
+		}
+	} else {
 		first = true
 		err = it.readFirstTimestamp(stream)
-	} else {
-		err = it.readNextTimestamp(stream)
 	}
+
 	if err != nil {
 		return false, false, err
 	}
@@ -99,7 +115,7 @@ func (it *TimestampIterator) ReadTimestamp(stream *encoding.IStream) (bool, bool
 // accordingly. It is exposed as a public method so that callers can control
 // the encoding / decoding of the time unit on their own if they choose.
 func (it *TimestampIterator) ReadTimeUnit(stream *encoding.IStream) error {
-	tuBits, err := stream.ReadByte()
+	tuBits, err := stream.ReadBits(8)
 	if err != nil {
 		return err
 	}
@@ -122,8 +138,14 @@ func (it *TimestampIterator) readFirstTimestamp(stream *encoding.IStream) error 
 	// NB(xichen): first time stamp is always normalized to nanoseconds.
 	nt := xtime.UnixNano(ntBits)
 	if it.TimeUnit == xtime.None {
-		it.TimeUnit = initialTimeUnit(nt, it.Opts.DefaultTimeUnit())
+		it.TimeUnit = initialTimeUnit(nt, it.defaultTimeUnit)
 	}
+
+	tes, ok := it.timeEncodingSchemes.SchemeForUnit(it.TimeUnit)
+	if !ok {
+		return errNoTimeSchemaForUnit
+	}
+	it.timeEncodingScheme = tes
 
 	err = it.readNextTimestamp(stream)
 	if err != nil {
@@ -147,30 +169,36 @@ func (it *TimestampIterator) readNextTimestamp(stream *encoding.IStream) error {
 
 // nolint: gocyclo
 func (it *TimestampIterator) tryReadMarker(stream *encoding.IStream) (time.Duration, bool, error) {
-	opcodeAndValue, success := it.tryPeekBits(stream, it.numBits)
-	if !success {
+	var (
+		numBits             = it.numBits
+		numValueBits        = it.numValueBits
+		opcodeAndValue, err = stream.PeekBits(numBits)
+	)
+
+	if err != nil {
 		return 0, false, nil
 	}
 
-	opcode := opcodeAndValue >> it.numValueBits
+	opcode := opcodeAndValue >> numValueBits
 	if opcode != it.markerEncodingScheme.Opcode() {
 		return 0, false, nil
 	}
 
 	var (
-		valueMask   = (1 << it.numValueBits) - 1
-		markerValue = int64(opcodeAndValue & uint64(valueMask))
+		valueMask   = (1 << numValueBits) - 1
+		markerValue = encoding.Marker(opcodeAndValue & uint64(valueMask))
 	)
-	switch encoding.Marker(markerValue) {
+
+	switch markerValue {
 	case it.markerEncodingScheme.EndOfStream():
-		_, err := stream.ReadBits(it.numBits)
+		_, err := stream.ReadBits(numBits)
 		if err != nil {
 			return 0, false, err
 		}
 		it.Done = true
 		return 0, true, nil
 	case it.markerEncodingScheme.Annotation():
-		_, err := stream.ReadBits(it.numBits)
+		_, err := stream.ReadBits(numBits)
 		if err != nil {
 			return 0, false, err
 		}
@@ -184,7 +212,7 @@ func (it *TimestampIterator) tryReadMarker(stream *encoding.IStream) (time.Durat
 		}
 		return markerOrDOD, true, nil
 	case it.markerEncodingScheme.TimeUnit():
-		_, err := stream.ReadBits(it.numBits)
+		_, err := stream.ReadBits(numBits)
 		if err != nil {
 			return 0, false, err
 		}
@@ -207,46 +235,27 @@ func (it *TimestampIterator) readMarkerOrDeltaOfDelta(
 ) (time.Duration, error) {
 	if !it.SkipMarkers {
 		dod, success, err := it.tryReadMarker(stream)
-		if err != nil {
-			return 0, err
-		}
-		if it.Done {
-			return 0, nil
-		}
-
-		if success {
-			return dod, nil
+		if success || err != nil || it.Done {
+			return dod, err
 		}
 	}
 
-	tes, exists := it.Opts.TimeEncodingSchemes().SchemeForUnit(it.TimeUnit)
-	if !exists {
-		return 0, fmt.Errorf("time encoding scheme for time unit %v doesn't exist", it.TimeUnit)
-	}
-
-	return it.readDeltaOfDelta(stream, tes)
+	return it.readDeltaOfDelta(stream)
 }
 
 func (it *TimestampIterator) readDeltaOfDelta(
 	stream *encoding.IStream,
-	tes encoding.TimeEncodingScheme,
 ) (time.Duration, error) {
 	if it.TimeUnitChanged {
-		// NB(xichen): if the time unit has changed, always read 64 bits as normalized
-		// dod in nanoseconds.
-		dodBits, err := stream.ReadBits(64)
-		if err != nil {
-			return 0, err
-		}
-
-		dod := encoding.SignExtend(dodBits, 64)
-		return time.Duration(dod), nil
+		return it.readFullTimestamp(stream)
 	}
 
 	cb, err := stream.ReadBits(1)
 	if err != nil {
 		return 0, err
 	}
+
+	tes := it.timeEncodingScheme
 	if cb == tes.ZeroBucket().Opcode() {
 		return 0, nil
 	}
@@ -289,6 +298,26 @@ func (it *TimestampIterator) readDeltaOfDelta(
 	return xtime.FromNormalizedDuration(dod, timeUnit), nil
 }
 
+func (it *TimestampIterator) readFullTimestamp(
+	stream *encoding.IStream,
+) (time.Duration, error) {
+	tes, exists := it.timeEncodingSchemes.SchemeForUnit(it.TimeUnit)
+	if !exists {
+		return 0, errNoTimeSchemaForUnit
+	}
+	it.timeEncodingScheme = tes
+	// NB(xichen): if the time unit has changed, always read 64 bits as normalized
+	// dod in nanoseconds.
+	dodBits, err := stream.ReadBits(64)
+	if err != nil {
+		return 0, err
+	}
+
+	dod := encoding.SignExtend(dodBits, 64)
+
+	return time.Duration(dod), nil
+}
+
 func (it *TimestampIterator) readAnnotation(stream *encoding.IStream) error {
 	antLen, err := it.readVarint(stream)
 	if err != nil {
@@ -298,7 +327,7 @@ func (it *TimestampIterator) readAnnotation(stream *encoding.IStream) error {
 	// NB: we add 1 here to offset the 1 we subtracted during encoding.
 	antLen = antLen + 1
 	if antLen <= 0 {
-		return fmt.Errorf("unexpected annotation length %d", antLen)
+		return errUnexpectedAnnotationLength
 	}
 
 	// TODO(xichen): use pool to allocate the buffer once the pool diff lands.
@@ -308,9 +337,7 @@ func (it *TimestampIterator) readAnnotation(stream *encoding.IStream) error {
 		return err
 	}
 	if n != antLen {
-		return fmt.Errorf(
-			"expected to read %d annotation bytes but read: %d",
-			antLen, n)
+		return errAnnotationTooFewBytes
 	}
 	it.PrevAnt = buf
 
@@ -320,12 +347,4 @@ func (it *TimestampIterator) readAnnotation(stream *encoding.IStream) error {
 func (it *TimestampIterator) readVarint(stream *encoding.IStream) (int, error) {
 	res, err := binary.ReadVarint(stream)
 	return int(res), err
-}
-
-func (it *TimestampIterator) tryPeekBits(stream *encoding.IStream, numBits uint8) (uint64, bool) {
-	res, err := stream.PeekBits(numBits)
-	if err != nil {
-		return 0, false
-	}
-	return res, true
 }