… src/dbnode/encoding/encoding_mock.go
… src/dbnode/encoding/scheme.go