<!--  Thanks for sending a pull request!  Here are some tips for you:
1. If this is your first time, please read our contributor guidelines: https://github.com/m3db/m3/blob/master/CONTRIBUTING.md and developer notes https://github.com/m3db/m3/blob/master/DEVELOPMENT.md
2. Please prefix the name of the pull request with the component you are updating in the format "[component] Change title" (for example "[dbnode] Support out of order writes") and also label this pull request according to what type of issue you are addressing. Furthermore, if this is a WIP or DRAFT PR, please create a draft PR instead: https://github.blog/2019-02-14-introducing-draft-pull-requests/
3. Ensure you have added or ran the appropriate tests for your PR: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#testing-changes
4. Follow the instructions for writing a changelog note: read more at https://github.com/m3db/m3/blob/master/DEVELOPMENT.md#updating-the-changelog
-->

**What this PR does / why we need it**:
<!--
If you have an issue this change addresses, please add the following details:
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->
Inline some really hot paths, and get rid of interface calls for marker/time encoding schemas, as those abstractions cost performance and are not in use anywhere outside the package.

```
name            old time/op  new time/op  delta
M3TSZDecode-12  42.9µs ± 0%  29.0µs ± 1%  -32.40%  (p=0.008 n=5+5)
M3TSZEncode-12  78.8µs ± 2%  67.3µs ± 0%  -14.62%  (p=0.008 n=5+5)
```

`read_data_files -B datapoints` benchmarks:

  | 10s tiles, old | 10s tiles, new | 1m tiles, old | 1m tiles, new
-- | -- | -- | -- | --
average | 9,056,942 | 11,997,466 | 10,406,413 | 14,105,973
stddev | 483,093 | 461,853 | 385,287 | 694,538



**Special notes for your reviewer**:

**Does this PR introduce a user-facing and/or backwards incompatible change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->
```release-note

```

**Does this PR require updating code package or user-facing documentation?**:
<!--
If no, just write "NONE" in the documentation-note block below.
If yes, describe which documentation you updated.
Note: Any changes that significantly updates how a code package is architectured or functions requires code package documentation updates in the form of updates to the README.md file in that package.
-->
```documentation-note

```