@vdarulis 2021-03-16
> @vdarulis FYI you can use the benchmarking mode of `read_data_files` utility (`-B datapoints`) to macro-benchmark a change like this on a large amount of real world data.
Thanks, added some data