@vdarulis 2021-03-19
> LGTM, left a few questions/suggestions inline.
> 
> As a general comment, I think test coverage of M3TSZ codec relies heavily on roundtrip checks. So my only concern is that when modifying both encoder and decoder at the same time, there is theoretic possibility to miss a breaking change (if it happens to be symmetric on encoder/decoder). And fiddling with bits has always been error prone (at least for me).
> 
> May I ask you to try reverting the changes on the encoder (can be done locally, no need to push) and then running the test suite with decoder changes only, for the peace of mind?

Sure thing. I actually did a pass on the encoder just because I could apply most of the learnings from decoder code.